// src/module/financial/financial-debt/financial-debt.enum.ts
export enum FINANCIAL_DEBT_TYPE_ENUM {
  LOAN = "LOAN",
  CREDIT_CARD = "CREDIT_CARD",
  MORTGAGE = "MORTGAGE",
  OTHER = "OTHER",
}

export enum FINANCIAL_DEBT_STATUS_ENUM {
  PENDING = "PENDING",
  ACTIVE = "ACTIVE",
  PAID_OFF = "PAID_OFF",
  CANCELLED = "CANCELLED",
  SETTLED = "SETTLED",
}

export enum FINANCIAL_DEBT_DIRECTION_ENUM {
  INCOMING = "INCOMING",
  OUTGOING = "OUTGOING",
}